Add files via upload
Origanal code. The program uses the ardounio-nano, a LCD which displays to you distance , a LED which repersents distance with brightness, and a ultra sonic sensor which measures distance. trig pin of the ultra sonic sensor goes to digital pin thirteen. echo pin of the ultra sonic sensor goes to digital pin 13, the Led pin is analog pin 0 and the LDC pins are rs = 12, en = 10, d4 = 5, d5 = 4, d6 = 3, d7 = 2.